+++ rmf_fleet_adapter/src/rmf_fleet_adapter/tasks/Delivery.cpp
@@ -21,42 +21,52 @@
 
 #include "Delivery.hpp"
 
+#include <rmf_ingestor_msgs/msg/ingestor_request_item.hpp>
+
 namespace rmf_fleet_adapter {
 namespace tasks {
 
 //==============================================================================
 std::shared_ptr<Task> make_delivery(
-    const rmf_task_msgs::msg::Delivery& request,
+    const rmf_task::ConstRequestPtr request,
     const agv::RobotContextPtr& context,
-    rmf_traffic::agv::Plan::Start pickup_start,
-    rmf_traffic::agv::Plan::Start dropoff_start)
+    const rmf_traffic::agv::Plan::Start pickup_start,
+    const rmf_traffic::Time deployment_time,
+    const rmf_task::agv::State finish_state)
 {
-  const auto& graph = context->navigation_graph();
 
-  const auto pickup_wp =
-      graph.find_waypoint(request.pickup_place_name)->index();
+  std::shared_ptr<const rmf_task::requests::DeliveryDescription> description =
+    std::dynamic_pointer_cast<
+      const rmf_task::requests::DeliveryDescription>(request->description());
+  
+  if (description == nullptr)
+    return nullptr;
 
   Task::PendingPhases phases;
   phases.push_back(
-        phases::GoToPlace::make(context, std::move(pickup_start), pickup_wp));
+    phases::GoToPlace::make(
+      context, std::move(pickup_start),
+      description->pickup_waypoint()));
 
   phases.push_back(
         std::make_unique<phases::DispenseItem::PendingPhase>(
           context,
-          request.task_id,
-          request.pickup_dispenser,
+          request->id(),
+          description->pickup_dispenser(),
           context->itinerary().description().owner(),
-          request.items));
+          description->items()));
 
-  const auto dropoff_wp =
-      graph.find_waypoint(request.dropoff_place_name)->index();
-
+  auto dropoff_start = description->dropoff_start(pickup_start);
   phases.push_back(
-        phases::GoToPlace::make(context, std::move(dropoff_start), dropoff_wp));
+    phases::GoToPlace::make(
+      context,
+      std::move(dropoff_start),
+      description->dropoff_waypoint()));
+
 
   std::vector<rmf_ingestor_msgs::msg::IngestorRequestItem> ingestor_items;
-  ingestor_items.reserve(request.items.size());
-  for(auto& dispenser_item : request.items){
+  ingestor_items.reserve(description->items().size());
+  for(const auto& dispenser_item : description->items()){
     rmf_ingestor_msgs::msg::IngestorRequestItem item{};
     item.type_guid = dispenser_item.type_guid;
     item.quantity = dispenser_item.quantity;
@@ -67,12 +77,18 @@ std::shared_ptr<Task> make_delivery(
   phases.push_back(
         std::make_unique<phases::IngestItem::PendingPhase>(
           context,
-          request.task_id,
-          request.dropoff_ingestor,
+          request->id(),
+          description->dropoff_ingestor(),
           context->itinerary().description().owner(),
           ingestor_items));
 
-  return Task::make(request.task_id, std::move(phases), context->worker());
+  return Task::make(
+    request->id(),
+    std::move(phases),
+    context->worker(),
+    deployment_time,
+    finish_state,
+    request);
 }
 
 } // namespace task